% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_edge_bundle.R
\name{geom_edge_bundle}
\alias{geom_edge_bundle}
\alias{geom_edge_bundle0}
\alias{geom_edge_bundle2}
\title{Create edge bundles between nodes}
\usage{
geom_edge_bundle(mapping = NULL, data = gCon(), position = "identity",
  arrow = NULL, lineend = "butt", show.legend = NA, n = 100,
  tension = 0.8, ...)

geom_edge_bundle2(mapping = NULL, data = gCon(), position = "identity",
  arrow = NULL, lineend = "butt", show.legend = NA, n = 100,
  tension = 0.8, ...)

geom_edge_bundle0(mapping = NULL, data = gCon(), position = "identity",
  arrow = NULL, lineend = "butt", show.legend = NA, tension = 0.8, ...)
}
\description{
The concept of edge bundles...
}
\references{
Holten, D. (2006). \emph{Hierarchical edge bundles: visualization
of adjacency relations in hierarchical data.} IEEE Transactions on
Visualization and Computer Graphics, \strong{12}(5), 741-748.
http://doi.org/10.1109/TVCG.2006.147
}

