% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_edge_density.R
\name{geom_edge_density}
\alias{geom_edge_density}
\title{Show edges as a density map}
\usage{
geom_edge_density(mapping = NULL, data = gEdges("short"),
  position = "identity", show.legend = NA, n = 100, ...)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2]{aes}}
or \code{\link[ggplot2]{aes_}}. By default x, y, xend and yend are mapped to
x, y, xend and yend in the edge data. Thus only edge_fill is really relevant
to set.}

\item{data}{The return of a call to \code{gEdges('short', ...)} or a
data.frame giving edges in the short format. See \code{\link{gEdges}} for
more details.}

\item{position}{Position adjustment, either as a string, or the result of a
call to a position adjustment function. Currently no meaningful position
adjustment exists for edges.}

\item{show.legend}{logical. Should this layer be included in the legends?
\code{NA}, the default, includes if any aesthetics are mapped. \code{FALSE}
never includes, and \code{TRUE} always includes.}

\item{n}{The number of points to estimate in the x and y direction, i.e. the
resolution of the raster.}

\item{...}{other arguments passed on to \code{\link[ggplot2]{layer}}. There
are three types of arguments you can use here:
\itemize{
 \item{Aesthetics: to set an aesthetic to a fixed value, like
 \code{color = "red"} or \code{size = 3.}}
 \item{Other arguments to the layer, for example you override the default
 \code{stat} associated with the layer.}
 \item{Other arguments passed on to the stat.}
}}
}
\description{
This geom makes it possible to add a layer showing edge presence as a density
map. Each edge is converted to \code{n} points along the line and a jitter is
applied. Based on this dataset a two-dimensional kernel density estimation is
applied and plotted as a raster image. The density is mapped to the alpha
level, making it possible to map a variable to the fill.
}
\note{
In order to avoid excessive typing edge aesthetic names are
automatically expanded. Because of this it is not necessary to write
\code{edge_colour} within the \code{aes()} call as \code{colour} will
automatically be renamed appropriately.
}
\section{Aesthetics}{

geom_edge_density understand the following aesthetics. Bold aesthetics are
automatically set, but can be overridden.
\itemize{
 \item{\strong{x}}
 \item{\strong{y}}
 \item{\strong{xend}}
 \item{\strong{yend}}
 \item{edge_fill}
 \item{filter}
}
}

\section{Computed variables}{


\describe{
 \item{x, y}{The coordinates for each pixel in the raster}
 \item{density}{The density associated with the pixel}
}
}
\examples{
require(igraph)
gr <- make_graph('bull')
E(gr)$class <- sample(letters[1:3], gsize(gr), replace = TRUE)

ggraph(gr, 'igraph', type = 'nicely') + geom_edge_density(aes(fill = class)) +
  geom_edge_link() + geom_node_point()

}
\author{
Thomas Lin Pedersen
}
\seealso{
Other geom_edge_*: \code{\link{geom_edge_diagonal}},
  \code{\link{geom_edge_elbow}},
  \code{\link{geom_edge_fan}},
  \code{\link{geom_edge_link}},
  \code{\link{geom_edge_loop}}
}

