% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_edge_sf.R
\name{geom_edge_sf}
\alias{geom_edge_sf}
\title{Draw edges as LINESTRINGs in geographical space}
\usage{
geom_edge_sf(
  mapping = NULL,
  data = get_sf_nodes(),
  stat = "sf",
  position = "identity",
  show.legend = NA,
  ...
)
}
\description{
Draw edges as LINESTRINGs in geographical space
}
