% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aaa.R, R/geom_axis_hive.R, R/geom_edge.R, R/geom_edge_arc.R, R/geom_edge_bundle.R, R/geom_edge_density.R, R/geom_edge_diagonal.R, R/geom_edge_elbow.R, R/geom_edge_fan.R, R/geom_edge_hive.R, R/geom_edge_link.R, R/geom_edge_loop.R, R/geom_treemap.R, R/ggproto-classes.R
\docType{data}
\name{StatFilter}
\alias{GeomAxisHive}
\alias{GeomEdgeBezier}
\alias{GeomEdgeBspline}
\alias{GeomEdgeDensity}
\alias{GeomEdgePath}
\alias{GeomEdgeSegment}
\alias{GeomTreemap}
\alias{StatAxisHive}
\alias{StatEdgeArc}
\alias{StatEdgeArc0}
\alias{StatEdgeArc2}
\alias{StatEdgeBundle}
\alias{StatEdgeBundle0}
\alias{StatEdgeBundle2}
\alias{StatEdgeDensity}
\alias{StatEdgeDiagonal}
\alias{StatEdgeDiagonal0}
\alias{StatEdgeDiagonal2}
\alias{StatEdgeElbow}
\alias{StatEdgeElbow0}
\alias{StatEdgeElbow2}
\alias{StatEdgeFan}
\alias{StatEdgeFan0}
\alias{StatEdgeFan2}
\alias{StatEdgeHive}
\alias{StatEdgeHive0}
\alias{StatEdgeHive2}
\alias{StatEdgeLink}
\alias{StatEdgeLink2}
\alias{StatEdgeLoop}
\alias{StatEdgeLoop0}
\alias{StatFilter}
\alias{ggraph-extensions}
\title{ggraph extensions to ggplot2}
\description{
This help page lists all exported ggproto classes defined by ggraph. In
general these should be of no concern to the user as the main interface to
the functionality is, as with ggplot2, the \code{geom_*} format. As opposed
to ggplot2 there really aren't any use for separate \code{stat_*} functions
as they are intimately linked to each geom and mixing and matching stats and
geoms would only cause a lot of trouble.
}
\details{
Many of the \code{geom_edge_*} geoms comes in different flavors dependent on
the functionality required. There will always be a base geom and some will
have a \code{geom_edge_*0} and \code{geom_edge_*2} version. The base geom
will, in the case of multiple versions, draw the edge as a sequence of small
segments. The different aesthetics will be repeated for each seqgment and a
counter will be added the enumerates the progression of segments, so that a
gradient of colour or size can be added along the edge by assigning the
respective aesthetic to \code{..index..}. \code{geom_edge_*2} will also draw
the edge as segments but will interpolate between the aesthetics at the end
points. This makes it possible for an edge to interpolate node properties of
its end nodes. \code{geom_edge_*2} is less performant than the base geom so
use only when interpolation is needed. \code{geom_edge_*0} is a
high-performance version that usually maps directly to a grid grob. It does
not allow for drawing gradients or interpolations though, so use only for
simple edge drawings.
}
\keyword{datasets}
\keyword{internal}

