% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph-package.R
\docType{package}
\name{ggraph-package}
\alias{ggraph-package}
\title{ggraph: Grammar of Graph Graphics}
\description{
There really should be some cool text and examples here. Will do once I
finish writting the cool code...
}

