% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_highschool.R
\docType{data}
\name{highschool}
\alias{highschool}
\title{Friendship among high school boys}
\format{
The graph is stored as an unnamed edgelist with a year attribute.
\describe{
\item{from}{The boy answering the question}
\item{to}{The boy being the answer to the question}
\item{year}{The year the friendship was reported}
}
}
\source{
Coleman, J. S. \emph{Introduction to Mathematical Sociology}. New York: Free
Press, pp.450-451.
}
\usage{
highschool
}
\description{
This dataset shows the friendship among high school boys as assessed by the
question: "What fellows here in school do you go around with most often?".
The question was posed twice, with one year in between (1957 and 1958) and
shows the evolution in friendship between the two timepoints.
}
\keyword{datasets}
