% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connections.R, R/edges.R
\name{internal_extractors}
\alias{internal_extractors}
\alias{collect_connections}
\alias{collect_edges}
\title{Internal data extractors}
\usage{
collect_connections(layout, from, to, ...)

collect_edges(layout)
}
\arguments{
\item{layout}{The layout data}

\item{from, to}{A numeric vector giving the indexes of the start and end nodes}

\item{...}{Additional parameters passed on to the specific method}
}
\description{
These functions exists for supporting different data structures. There is no
need to call these directly
}
\keyword{internal}
