% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_dendrogram.R
\name{layout_dendrogram_even}
\alias{layout_dendrogram_even}
\title{Even layout for layout_dendrogram}
\usage{
layout_dendrogram_even(graph, ...)
}
\arguments{
\item{graph}{A dendrogram object}

\item{...}{parameters passed on to \code{\link{layout_dendrogram_dendrogram}}}
}
\value{
A data.frame as \code{\link{layout_dendrogram_dendrogram}}
}
\description{
This layout sets the heights of the branch points to be the maximum distance
to a leaf. In this way the branch points are spread out evenly in the
y-direction. After modifying the height it calls
\code{\link{layout_dendrogram_dendrogram}}.
}
\note{
This function is not intended to be used directly but by setting
\code{layout = 'even'} in \code{\link{create_layout}}
}
