% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_igraph.R
\name{layout_igraph_igraph}
\alias{layout_igraph_igraph}
\title{Use igraph layout algorithms for layout_igraph}
\usage{
layout_igraph_igraph(graph, type, circular, offset = pi/2,
  use.dummy = FALSE, ...)
}
\arguments{
\item{graph}{An igraph object.}

\item{type}{The type of layout algorithm to apply. See
\code{\link[igraph]{layout_}} for links to the layouts supplied by igraph.}

\item{circular}{Logical. Should the layout be transformed to a circular
representation. Defaults to \code{FALSE}. Only applicable to
\code{type = 'tree'} and \code{type = 'sugiyama'}.}

\item{offset}{If \code{circular = TRUE}, where should it begin. Defaults to
\code{pi/2} which is equivalent to 12 o'clock.}

\item{use.dummy}{Logical. In the case of \code{type = 'sugiyama'} should the
dummy-infused graph be used rather than the original. Defaults to
\code{FALSE}.}

\item{...}{Arguments passed on to the respective layout functions}
}
\value{
A data.frame with the columns \code{x}, \code{y}, \code{circular} as
well as any information stored as vertex attributes on the igraph object.
}
\description{
This layout function makes it easy to apply one of the layout algorithms
supplied in igraph when plotting with ggraph. Layout names are auto completed
so there is no need to write \code{layout_with_graphopt} or
\code{layout_as_tree}, just \code{graphopt} and \code{tree} (though the
former will also work if you want to be super explicit). Circular layout is
only supported for tree-like layout (\code{tree} and \code{sugiyama}) and
will throw an error when applied to other layouts.
}
\note{
This function is not intended to be used directly but by setting
\code{layout = 'igraph'} in \code{\link{createLayout}}
}

