% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_igraph.R
\name{layout_igraph_manual}
\alias{layout_igraph_manual}
\title{Manually specify a layout for layout_igraph}
\usage{
layout_igraph_manual(graph, node.positions, circular)
}
\arguments{
\item{graph}{An igraph object}

\item{node.positions}{A data.frame with the columns \code{x} and \code{y}
(additional columns are ignored).}

\item{circular}{Ignored}
}
\value{
A data.frame with the columns \code{x}, \code{y}, \code{circular} as
well as any information stored as vertex attributes on the igraph object.
}
\description{
This layout function lets you pass the node positions in manually. Each row
in the supplied data frame will correspond to a vertex in the igraph object
matched by index.
}
\seealso{
Other layout_igraph_*: \code{\link{layout_igraph_auto}},
  \code{\link{layout_igraph_circlepack}},
  \code{\link{layout_igraph_dendrogram}},
  \code{\link{layout_igraph_hive}},
  \code{\link{layout_igraph_linear}},
  \code{\link{layout_igraph_partition}},
  \code{\link{layout_igraph_treemap}}
}
