% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_tbl_graph.R
\name{layout_tbl_graph_auto}
\alias{layout_tbl_graph_auto}
\title{Automatically pick a layout based on graph type}
\usage{
layout_tbl_graph_auto(graph, circular, ...)
}
\arguments{
\item{graph}{A tbl_graph object}

\item{circular}{Logical. Should the layout be transformed to a circular
representation. Defaults to \code{FALSE}. Only applicable to
\code{algorithm = 'tree'} and \code{algorithm = 'sugiyama'}.}

\item{...}{Arguments passed on to the chosen layout}
}
\value{
A data.frame with the columns \code{x}, \code{y}, \code{circular} as
well as any information stored as node variables in the tbl_graph object.
}
\description{
This function infers the layout from the graph structure and is the default
when calling \code{\link{ggraph}}. If the graph is either a tree or a forest
the layout will be `dendrogram` if the nodes contains a height variable or
`tree` if not. If the tree is a DAG the `sygiyama` layout will be used.
Otherwise the `nicely` layout will be used.
}
\seealso{
Other layout_tbl_graph_*: \code{\link{layout_tbl_graph_circlepack}},
  \code{\link{layout_tbl_graph_dendrogram}},
  \code{\link{layout_tbl_graph_hive}},
  \code{\link{layout_tbl_graph_igraph}},
  \code{\link{layout_tbl_graph_linear}},
  \code{\link{layout_tbl_graph_manual}},
  \code{\link{layout_tbl_graph_partition}},
  \code{\link{layout_tbl_graph_treemap}}
}
\concept{layout_tbl_graph_*}
