% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_tbl_graph.R
\name{layout_tbl_graph_manual}
\alias{layout_tbl_graph_manual}
\title{Manually specify a layout for layout_tbl_graph}
\usage{
layout_tbl_graph_manual(graph, x, y, circular)
}
\arguments{
\item{graph}{An igraph object}

\item{circular}{Ignored}

\item{node.positions}{A data.frame with the columns \code{x} and \code{y}
(additional columns are ignored).}
}
\value{
A data.frame with the columns \code{x}, \code{y}, \code{circular} as
well as any information stored as node variables in the tbl_graph.
}
\description{
This layout function lets you pass the node positions in manually. The
supplied positions must match the order of the nodes in the tbl_graph
}
\seealso{
Other layout_tbl_graph_*: \code{\link{layout_tbl_graph_auto}},
  \code{\link{layout_tbl_graph_circlepack}},
  \code{\link{layout_tbl_graph_dendrogram}},
  \code{\link{layout_tbl_graph_hive}},
  \code{\link{layout_tbl_graph_igraph}},
  \code{\link{layout_tbl_graph_linear}},
  \code{\link{layout_tbl_graph_partition}},
  \code{\link{layout_tbl_graph_treemap}}
}
\concept{layout_tbl_graph_*}
