% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_graph.R
\name{layout_to_table}
\alias{layout_to_table}
\title{Convert a layout to a table}
\usage{
layout_to_table(layout, graph, ...)
}
\arguments{
\item{layout}{A supported object}

\item{graph}{A \code{tbl_graph}}

\item{...}{passed on to implementations}
}
\value{
A valid data.frame
}
\description{
This generic takes care of dispatching various layout types (names,
functions, tables) to their respective functions that will return a valid
layout table.
}
\keyword{internal}
