% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cappedPath.R
\name{makeContent.cappedpathgrob}
\alias{makeContent.cappedpathgrob}
\title{Dynamic capping of paths}
\usage{
\method{makeContent}{cappedpathgrob}(x)
}
\description{
This function takes care of updating which parts of the paths get removed
when the device dimensions are updated.
}
\keyword{internal}
