% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textAlong.R
\name{makeContent.textalong}
\alias{makeContent.textalong}
\title{Text angeled acording to line}
\usage{
\method{makeContent}{textalong}(x)
}
\description{
This function takes care of recalculating the angel of the text as the device
size changes
}
\keyword{internal}
