% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_sec.R
\name{network_to_igraph}
\alias{network_to_igraph}
\title{Coerce network to igraph}
\usage{
network_to_igraph(graph)
}
\arguments{
\item{graph}{A graph as a network object}
}
\value{
A representation of the same graph as given in the function call but
as an igraph object.
}
\description{
This utility function performs a conversion of network objects from the
network package into igraph object compatible with ggraph. Edge and node
attributes are preserved which, for the context of ggraph, is the most
important.
}
