% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\docType{import}
\name{reexports}
\alias{reexports}
\alias{scale_color_viridis}
\alias{scale_fill_viridis}
\title{Objects exported from other packages}
\keyword{internal}
\description{
These objects are imported from other packages. Follow the links
below to see their documentation.

\describe{
  \item{viridis}{\code{\link[viridis:scale_viridis]{scale_color_viridis}}, \code{\link[viridis:scale_viridis]{scale_fill_viridis}}}
}}

