% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geometry.R
\name{scale_type.geometry}
\alias{scale_type.geometry}
\title{Define default scale type for geometry}
\usage{
\method{scale_type}{geometry}(x)
}
\description{
This function is quite useless as geometry is not meant to be scaled, but it
is a requirement for ggplot2 to handle it correctly.
}
\keyword{internal}
