% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_whigs.R
\docType{data}
\name{whigs}
\alias{whigs}
\title{Membership network of American Whigs}
\format{
The data is stored as an incidence matrix with persons as rows and
organizations as columns. A 0 means no membership while a one means
membership.
}
\source{
\url{https://github.com/kjhealy/revere/blob/master/data/PaulRevereAppD.csv}
adapted from:

Fischer, David H. (1995) \emph{Paul Revere's Ride}. Oxford University Press
}
\usage{
whigs
}
\description{
This dataset shows the membership of 136 colonial Americans in 5 whig
organization and is a bipartite graph. The data appeared in the appendix to
David Hackett Fischer's \emph{Paul Revere's Ride} (Oxford University Press,
1995) and compiled by Kieran Healy for the blog post
\href{https://kieranhealy.org/blog/archives/2013/06/09/using-metadata-to-find-paul-revere/}{Using Metadata to Find Paul Revere}.
}
\keyword{datasets}
