% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_edge_sf.R
\name{geom_edge_sf}
\alias{geom_edge_sf}
\title{Draw edges as LINESTRINGs in geographical space}
\usage{
geom_edge_sf(
  mapping = NULL,
  data = get_sf_edges(),
  stat = "sf",
  position = "identity",
  show.legend = NA,
  ...
)
}
\arguments{
\item{mapping}{Set of aesthetic mappings created by \code{\link[ggplot2:aes]{ggplot2::aes()}}
or \code{\link[ggplot2:aes_]{ggplot2::aes_()}}. By default geometry is mapped to the geometry in
the edge data.}
}
\description{
This geom is equivalent in functionality to \code{\link[ggplot2:ggsf]{ggplot2::geom_sf()}} for \code{LINESTRING}
geometries and allows for plotting of edges in their geographical space in
different colours, linetypes and widths.
}
\section{Aesthetics}{

\code{geom_edge_sf} understand the following aesthetics.
\itemize{
\item alpha
\item colour
\item linetype
}
}

\examples{
if (require("sfnetworks", quietly = TRUE)) {
  gr <- as_sfnetwork(roxel)
  ggraph(gr, 'sf') + geom_edge_sf()
}

}
\seealso{
Other geom_edge_*: 
\code{\link{geom_edge_arc}()},
\code{\link{geom_edge_bend}()},
\code{\link{geom_edge_density}()},
\code{\link{geom_edge_diagonal}()},
\code{\link{geom_edge_elbow}()},
\code{\link{geom_edge_fan}()},
\code{\link{geom_edge_hive}()},
\code{\link{geom_edge_link}()},
\code{\link{geom_edge_loop}()},
\code{\link{geom_edge_parallel}()},
\code{\link{geom_edge_point}()},
\code{\link{geom_edge_span}()},
\code{\link{geom_edge_tile}()}
}
\author{
Lorena Abad
}
\concept{geom_edge_*}
