% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph-package.R
\docType{package}
\name{ggraph-package}
\alias{ggraph-package}
\title{ggraph: Grammar of Graph Graphics}
\description{
This package is an extension of the grammar of graphics defined in
\code{\link{ggplot2}} for relational data such as networks and trees.
\code{ggraph} adds a long list of new geoms and scales as well as a new plot
constructor that helps you build up your network visualization layer by layer
in the same manner as you know from \code{ggplot2}.
}
\seealso{
Useful links:
\describe{
  \item{Development version}{\url{https://github.com/thomasp85/ggraph}}
  \item{Feature requests and bug reports}{\url{https://github.com/thomasp85/ggraph/issues}}
}
}
\author{
Thomas Lin Pedersen \email{thomasp85@gmail.com}
}
