% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggraph-package.R
\docType{package}
\name{ggraph-package}
\alias{ggraph-package}
\alias{_PACKAGE}
\title{ggraph: An Implementation of Grammar of Graphics for Graphs and Networks}
\description{
\if{html}{\figure{logo.png}{options: align='right' alt='logo' width='120'}}

The grammar of graphics as implemented in ggplot2 is a poor fit for
    graph and network visualizations due to its reliance on tabular data input.
    ggraph is an extension of the ggplot2 API tailored to graph visualizations
    and provides the same flexible approach to building up plots layer by layer.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://ggraph.data-imaginist.com}
  \item \url{https://github.com/thomasp85/ggraph}
  \item Report bugs at \url{https://github.com/thomasp85/ggraph/issues}
}

}
\author{
\strong{Maintainer}: Thomas Lin Pedersen \email{thomasp85@gmail.com} (\href{https://orcid.org/0000-0002-5147-4711}{ORCID})

Other contributors:
\itemize{
  \item RStudio [copyright holder]
}

}
