% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_colourbar.R, R/edge_direction.R
\name{guide_train.edge_colourbar}
\alias{guide_train.edge_colourbar}
\alias{guide_train.edge_direction}
\alias{guide_merge.edge_direction}
\alias{guide_geom.edge_direction}
\alias{guide_gengrob.edge_direction}
\title{Helper methods for guides}
\usage{
\method{guide_train}{edge_colourbar}(guide, scale, aesthetic = NULL)

\method{guide_train}{edge_direction}(guide, scale, aesthetic = NULL)

\method{guide_merge}{edge_direction}(guide, new_guide)

\method{guide_geom}{edge_direction}(guide, layers, default_mapping)

\method{guide_gengrob}{edge_direction}(guide, theme)
}
\description{
Helper methods for guides
}
\keyword{internal}
