% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/edge_colourbar.R
\name{guide_edge_coloursteps}
\alias{guide_edge_coloursteps}
\alias{guide_edge_colorsteps}
\title{Coloursteps legend for edges}
\usage{
guide_edge_coloursteps(
  even.steps = TRUE,
  show.limits = NULL,
  ticks = FALSE,
  ...,
  available_aes = c("edge_colour", "edge_fill")
)

guide_edge_colorsteps(
  even.steps = TRUE,
  show.limits = NULL,
  ticks = FALSE,
  ...,
  available_aes = c("edge_colour", "edge_fill")
)
}
\arguments{
\item{even.steps}{Should the rendered size of the bins be equal, or should
they be proportional to their length in the data space? Defaults to \code{TRUE}}

\item{show.limits}{Logical. Should the limits of the scale be shown with
labels and ticks. Default is \code{NULL} meaning it will take the value from the
scale. This argument is ignored if \code{labels} is given as a vector of
values. If one or both of the limits is also given in \code{breaks} it will be
shown irrespective of the value of \code{show.limits}.}

\item{ticks}{A logical specifying if tick marks on the colourbar should be
visible.}

\item{...}{
  Arguments passed on to \code{\link[ggplot2:guide_coloursteps]{ggplot2::guide_coloursteps}}
  \describe{
    \item{\code{}}{}
  }}

\item{available_aes}{A vector of character strings listing the aesthetics
for which a colourbar can be drawn.}
}
\value{
A guide object
}
\description{
This function is equivalent to \code{\link[ggplot2:guide_coloursteps]{ggplot2::guide_coloursteps()}} but
works for edge aesthetics.
}
