% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout_manual.R
\name{layout_tbl_graph_manual}
\alias{layout_tbl_graph_manual}
\title{Manually specify a layout for layout_tbl_graph}
\usage{
layout_tbl_graph_manual(graph, x, y, circular)
}
\arguments{
\item{graph}{An \code{tbl_graph} object}

\item{x, y}{Expressions with the x and y positions of the nodes}

\item{circular}{Ignored}
}
\value{
A data.frame with the columns \code{x}, \code{y}, \code{circular} as
well as any information stored as node variables in the tbl_graph.
}
\description{
This layout function lets you pass the node positions in manually. The
supplied positions must match the order of the nodes in the tbl_graph
}
\seealso{
Other layout_tbl_graph_*: 
\code{\link{layout_tbl_graph_auto}()},
\code{\link{layout_tbl_graph_backbone}()},
\code{\link{layout_tbl_graph_centrality}()},
\code{\link{layout_tbl_graph_circlepack}()},
\code{\link{layout_tbl_graph_dendrogram}()},
\code{\link{layout_tbl_graph_eigen}()},
\code{\link{layout_tbl_graph_fabric}()},
\code{\link{layout_tbl_graph_focus}()},
\code{\link{layout_tbl_graph_hive}()},
\code{\link{layout_tbl_graph_igraph}()},
\code{\link{layout_tbl_graph_linear}()},
\code{\link{layout_tbl_graph_matrix}()},
\code{\link{layout_tbl_graph_partition}()},
\code{\link{layout_tbl_graph_pmds}()},
\code{\link{layout_tbl_graph_stress}()},
\code{\link{layout_tbl_graph_treemap}()},
\code{\link{layout_tbl_graph_unrooted}()}
}
\concept{layout_tbl_graph_*}
